+++ OnlineRailwayReservation/src/main/java/com/sunbeam/entity/Station.java
@@ -20,7 +20,7 @@ public class Station {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "station_id")
-	private int id;
+	private Long id;
 
 	@Column(name = "station_name")
 	private String stationName;
@@ -34,11 +34,11 @@ public class Station {
 	@OneToOne(mappedBy = "destinationStation", fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
 	private Route routeDestination;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public Station(int id, String stationName, Route routeSource, Route routeDestination) {
+	public Station(Long id, String stationName, Route routeSource, Route routeDestination) {
 		this.id = id;
 		this.stationName = stationName;
 		this.routeSource = routeSource;
@@ -48,7 +48,7 @@ public class Station {
 	public Station() {
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
